+++ Backend/src/test/java/com/luckyseven/backend/domain/team/service/TeamServiceTest.java
@@ -4,12 +4,12 @@ import com.luckyseven.backend.domain.team.dto.TeamCreateRequest;
 import com.luckyseven.backend.domain.team.dto.TeamCreateResponse;
 import com.luckyseven.backend.domain.team.dto.TeamDashboardResponse;
 import com.luckyseven.backend.domain.team.dto.TeamJoinResponse;
-import com.luckyseven.backend.domain.team.entity.Budget;
+import com.luckyseven.backend.domain.budget.entity.Budget;
 import com.luckyseven.backend.domain.team.entity.Expense;
 import com.luckyseven.backend.domain.member.entity.Member;
 import com.luckyseven.backend.domain.team.entity.Team;
 import com.luckyseven.backend.domain.team.entity.TeamMember;
-import com.luckyseven.backend.domain.team.repository.BudgetRepository;
+import com.luckyseven.backend.domain.budget.dao.BudgetRepository;
 import com.luckyseven.backend.domain.team.repository.ExpenseRepository;
 import com.luckyseven.backend.domain.team.repository.TeamMemberRepository;
 import com.luckyseven.backend.domain.team.repository.TeamRepository;
